Updated Spine Class
Added functionality of the spine class to get the velocity, acceleration
of every node at any time t.

TODO:

Use ode45 to integrate these derivatives.

Note: ODE45 is usually used between two time steps from the beginning to the end of the simulation, however, the way this code has been structured, the derivatives come from the properties of each vertebrate object in the spine, so these need updating at each time step. This would not be done if we were to use the ode45 in its ordinary usage.

The issue that requries solving, or demonstrating that it is not an issue, is that when using ode45 and passing two time steps, MATLAB assumes that it is desired to use intermediate steps in this integral, however for this case we dont want to do this. We want the two time steps passed to be interpreted as t, t+dt. i.e we do one step of the integration at a time, then after doing one step, we reassign the values of displacement, velocity to each of the vertebrae, so that the next integration step is referencing the correct values of displacement, velocity.

Hopefully, passing through a small enough time step to the ode45 function will allow it to use zero intermediate steps.